+++ postinstall.js
@@ -9,14 +9,35 @@ var https = require('https');
 const execSync = require('child_process').execSync;
 
 // How to get to the MQ Redistributable Client package
-var protocol = "https://"
+var protocol = "https://";
 var host="public.dhe.ibm.com";
 var baseDir="ibmdl/export/pub/software/websphere/messaging/mqdev";
 var redistDir=baseDir+"/redist";
 var macDir=baseDir+"/mactoolkit";
-var vrm="9.2.0";
+
+// This is the version (VRM) of MQ associated with this level of package
+var vrm="9.2.1";
+
+// Allow overriding the VRM - but you need to be careful as this package
+// may depend on MQI features in the listed version. Must be given in the
+// same format eg "1.2.3". Note that IBM keeps a limited set of versions for
+// download - once a version of MQ is no longer supported, that level of
+// Redistributable Client package is removed from public sites.
+var vrmenv=process.env['MQIJS_VRM'];
+if (vrmenv != null) {
+  vrm=vrmenv;
+}
+
+// Set the fixpack which is also part of the downloadable name. Default
+// to 0, which is the only allowable level for CD versions of MQ. Only
+// the LTS level gets fixpacks.
 var vrmf=vrm + ".0";
-var macVrmf="9.2.0.0" 
+// Allow overriding the fixpack if VRM indicates an LTS version
+var fixpack=process.env['MQIJS_FIXPACK'];
+if (fixpack != null && vrm.endsWith('.0')) {
+  vrmf=vrm+"."+fixpack;
+}
+
 var file=vrmf + "-IBM-MQC-Redist-"; // will be completed by platform filetype
 var redistTitle="IBM MQ Redistributable C Client";
 
@@ -29,6 +50,7 @@ var unpackCommand;
 var unwantedDirs;
 const newBaseDir="redist";
 var currentPlatform=process.platform;
+var preferredVersion=14;
 //currentPlatform='darwin'; // for forcing a platform test
 
 // Some functions used at the end of the operation
@@ -90,6 +112,7 @@ function removePattern(d,type) {
 
 // Remove directories from the client that are not needed for Node execution
 function removeUnneeded() {
+  var d;
   var doNotRemove = process.env['MQIJS_NOREMOVE'];
   if (doNotRemove != null) {
     console.log('Environment variable set to keep all files in client package');
@@ -104,14 +127,14 @@ function removeUnneeded() {
     }
 
     if (currentPlatform === 'win32') {
-      var d = path.join(newBaseDir,"bin64");
+      d = path.join(newBaseDir,"bin64");
       removePattern(d,/.exe$/);
       removePattern(d,/^imq.*.dll$/);
     } else {
-      var d= path.join(newBaseDir,"lib");
+      d= path.join(newBaseDir,"lib");
       removePattern(d,/lib.*so/);
       removePattern(d,/amqtrc.fmt/);
-      var d= path.join(newBaseDir,"lib64");
+      d= path.join(newBaseDir,"lib64");
       removePattern(d,/libimq.*so/);
       removePattern(d,/libedit.so/);
     }
@@ -119,6 +142,21 @@ function removeUnneeded() {
 
   cleanup();
 }
+
+function checkNodeVersion() {
+  // V10/V12 seem to have unexpected internal failures in some levels,
+  // possibly related to the ffi-napi package. But V14
+  // looks to be OK. Rather than enforce that as a prereq level, at least
+  // give a warning to point at V14 as a preferred version.
+  var major=process.versions.node.split('.')[0];
+  if (major < preferredVersion) {
+    console.warn("Current Node version is " + process.versions.node);
+    console.warn("Preferred version is at least " + preferredVersion);
+  }
+}
+
+// Start main processing here.
+checkNodeVersion();
 
 // If a particular environment variable is set, do not try to install
 // the Redist client package. I did consider doing this automatically by
@@ -130,7 +168,7 @@ if (doit != null) {
   process.exit(0);
 }
 
-// Start main processing here. Check if the install is for an environment
+//Check if the install is for an environment
 // where there is a Redistributable Client.
 if (currentPlatform === 'win32') {
   file=file+"Win64.zip";
@@ -147,12 +185,12 @@ if (currentPlatform === 'win32') {
 //
 //  dir=macDir
 //  title="IBM MacOS Toolkit for Developers"
-//  file=macVrmf + "-IBM-MQ-Toolkit-MacX64" + ".tar.gz"
+//  file=vrmf + "-IBM-MQ-Toolkit-MacX64" + ".tar.gz"
 //  unpackCommand="mkdir -p " +  newBaseDir + " && tar -xvzf " + file + " -C " + newBaseDir;
 //  unwantedDirs=[ "samp", "bin","inc","java", "gskit8/lib", ".github" ];
 } else {
   console.log("No redistributable client package available for this platform.");
-  console.log("If an MQ Client library exists for the platform, install it manually.")
+  console.log("If an MQ Client library exists for the platform, install it manually.");
   process.exit(0);
 }